requirements.txt
Flask
pure-python-adb
Pillow
